requirements.txt
Flask
pymysql
psycopg2-binary
python-dotenv
APScheduler
gunicorn
